# api/rest/docs/swagger.yaml
definitions:
...
  rest.DemandPartnerOptimizationUpdateRequest:
    properties:
      browser:
        type: string
      country:
        type: string
      demand_partners:
        type: string
      device_type:
        type: string
      domain:
        type: string
      factor:
        type: number
      os:
        type: string
      placement_type:
        type: string
      publisher:
        type: string
    type: object
  rest.DemandPartnerOptimizationUpdateResponse:
    properties:
      rule_id:
        type: string
      status:
        description: 'in: body'
        type: string
    type: object
...
  rest.MetadataUpdateRequest:
    properties:
      data: {}
      key:
        type: string
      version:
        type: string
    type: object
  rest.MetadataUpdateRespose:
    properties:
      status:
        description: 'in: body'
        type: string
      transaction_id:
        type: string
    type: object
  rest.PriceFactorUpdateRequest:
    properties:
      country:
        type: string
      domain:
        type: string
      factor:
        type: number
      mobile:
        type: boolean
      publisher:
        type: string
    type: object
  rest.PriceFactorUpdateRespose:
    properties:
      status:
        description: 'in: body'
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /dpo/delete:
    delete:
      description: Delete demand partner optimization rule for publisher.
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      tags:
      - dpo
  /dpo/get:
    get:
      consumes:
      - application/json
      description: Get demand partner optimization rules for publisher.
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      tags:
      - dpo
  /dpo/set:
    post:
      consumes:
      - application/json
      description: Update demand partner optimization rule for a publisher.
      parameters:
      - description: Demand Partner Optimization update rule
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.DemandPartnerOptimizationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.DemandPartnerOptimizationUpdateResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - dpo
  /metadata/update:
    post:
      consumes:
      - application/json
      description: Update new bidder metadata.
      parameters:
      - description: Metadata update Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.MetadataUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MetadataUpdateRespose'
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
  /price/factor:
    post:
      consumes:
      - application/json
      description: Update new bidder price factor.
      parameters:
      - description: PriceFactor update Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.PriceFactorUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.PriceFactorUpdateRespose'
      security:
      - ApiKeyAuth: []
      tags:
      - metadata
...
swagger: "2.0"